A lot.
Disabled HUD elements and added in a custom weapon switching system.
Reduced the number of flies.
Made the map generation system use a set of predefined colors for key/door pairs based on the quadrant of the map.
The game must now be restart (as in, loading a new map) to generate a new puzzle (to avoid crashing).
Disabled Suit Zoom and Walk keys.
Disabled the flashlight sound.
Some other misc changes to the map generation system.
Added back in the florescent light entity with some changes.
Reduced the sound level of flies.
Changed the messages and style of keys to recognize doors, even when they're not the key's target door.
Fixed a bug with The Weeping Gman where he may attack the last spot where the player was before they died.
Modified the pausing system to track if The Weeping Gman was still frozen when the pause period had ended, and related tweaks.
Tweaked the hearing system of the Weeping Gman.
Fixed a bug with The Weeping Gman where his walk speed would be his run speed instead.
Modified the "retry cnav" system to delay more slowly at first.
The Weeping Gman will now want to run if his instability is over 50%.
The Weeping Gman will now give himself a little moving space before attempting to open a door.
Changed the search system of The Weeping Gman to use roulette selection.
Moved the cyan debug line (the one for target_last_known_position) outside of a bogus requirement (I guess? I dunno).
Tweaked the Weeping Gman's instability system to hint more clearly.
Tweaked the Weeping Gman's instability system to tick the timer more frequently when increasing and less frequently when decreasing instability.
Added a Cassette Player SWEP. The player can now play preset music while they try not to die.
Added a Map SWEP. Currently it loads a texture based on the name of the current map.
Renamed sent_tv_voltage_tester to sent_tv_voltagetester.
Updated the output map to fix the loud doors and to include the florescent lights.
Updated the FGD to include the florescent light entity.