OUT OF ALPHA, INTO BETA!
Reduced the delay for the intro from 10 sec to 1.25.
Fixed the bugs with the intro sequence.
Modified the intro sequence to take into account if the game is in HDR vs LDR.
The game now records when the player most recently played to determine if the intro should play or not (1 hour cooldown).
The game no longer has the sped-up simulation period at the start.
Added a new entity for dealing with ending the game, sent_tv_end_game (because other methods were not working weirdly).
Modified Cable IO entities to update their render data at a random interval between 3 and 6.
Made power and system IO entities not render their model when not in debug mode.
TWG now has a lower light threshold.
Updated tv_bangclaw. Now supports HDR. Fixed a few problems and got it fully functional!
Removed the other bsps (but left the vmfs for them).
Modified the telephone model's texture to make it clear that the phone can be interacted with, not just the dial.
Updated all the fuse materials to put the phongfresnelranges in quotes (minor change really).
Added the new entity, sent_tv_end_game, to the FGD.
Other misc. changes.